Cleaned up project spec, made generated page
Now the index.html file is generated from a list of projects described
as object data in `projects.js`, and descriptions in markdown files
within the `/projects` directory.

Running `npm start` will run a small node program to read in the
projects, parse the markdown, and merge it with the index template file
to generate the final files in the `/dist` folder.